refactor: restructure code
https://github.com/Attacktive/troubleshooter-editor-back-end/issues/58
