+++ src/main/kotlin/com/github/attacktive/troubleshootereditor/common/diff/DiffResult.kt
@@ -1,0 +1,6 @@
+package com.github.attacktive.troubleshootereditor.common.diff
+
+interface DiffResult {
+	val properties: PropertiesDiffResult?
+	val hasChanges: Boolean
+}
