+++ src/main/kotlin/com/github/attacktive/troubleshootereditor/common/diff/PropertiesDiffResult.kt
@@ -1,3 +1,3 @@
-package com.github.attacktive.troubleshootereditor.model
+package com.github.attacktive.troubleshootereditor.common.diff
 
 data class PropertiesDiffResult(val inserts: List<Pair<String, String>>, val updates: List<Pair<String, String>>, val deletes: List<String>)
