+++ src/main/kotlin/com/github/attacktive/troubleshootereditor/common/extension/Extensions.kt
@@ -1,6 +1,6 @@
-package com.github.attacktive.troubleshootereditor.extension
+package com.github.attacktive.troubleshootereditor.common.extension
 
-import com.github.attacktive.troubleshootereditor.model.Identifiable
+import com.github.attacktive.troubleshootereditor.module.Identifiable
 
 fun <I, T: Identifiable<I>> MutableSet<T>.findById(id: I): T? {
 	return stream()
