--- src/main/kotlin/com/github/attacktive/troubleshootereditor/model/DiffResult.kt
@@ -1,10 +1,0 @@
-package com.github.attacktive.troubleshootereditor.model
-
-import java.sql.Connection
-import java.sql.PreparedStatement
-
-interface DiffResult {
-	val hasChanges: Boolean
-
-	fun getStatements(connection: Connection): MutableList<PreparedStatement>
-}
