--- src/main/kotlin/com/github/attacktive/troubleshootereditor/model/Identifiable.kt
@@ -1,5 +1,0 @@
-package com.github.attacktive.troubleshootereditor.model
-
-interface Identifiable<T> {
-	fun getId(): T
-}
