+++ src/main/kotlin/com/github/attacktive/troubleshootereditor/module/Identifiable.kt
@@ -1,0 +1,5 @@
+package com.github.attacktive.troubleshootereditor.module
+
+interface Identifiable<T> {
+	fun getId(): T
+}
