+++ src/main/kotlin/com/github/attacktive/troubleshootereditor/upload/UploadService.kt
@@ -3,7 +3,7 @@ package com.github.attacktive.troubleshootereditor.upload
 import java.io.File
 import java.io.FileOutputStream
 import java.util.UUID
-import com.github.attacktive.troubleshootereditor.configuration.PropertiesConfiguration
+import com.github.attacktive.troubleshootereditor.common.configuration.PropertiesConfiguration
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
 import org.springframework.web.multipart.MultipartFile
